Fixes memory leak

+++ yolo/tools/solver.py
@@ -96,7 +96,7 @@ class ModelTrainer:
             for loss_name, loss_val in loss_each.items():
                 if self.use_ddp:  # collecting loss for each batch
                     distributed.all_reduce(loss_val, op=distributed.ReduceOp.AVG)
-                total_loss[loss_name] += loss_val * batch_size
+                total_loss[loss_name] += float(loss_val * batch_size)
             total_samples += batch_size
             self.progress.one_batch(loss_each)
